Split gossamer's distillate-facing givens into a component
`gossamer.core` has fan-in 44, so everything beneath it is compiled by
most of the tree — and its dependency on `distillate` rested on five
givens in one file: `textDecodable`, which decodes `Text` from bytes
through the `CharDecoder` in scope, and the four `enumIdentification`
instances that name enum cases in kebab, snake, pascal and camel case.
Nothing else in the component references a distillate name.

They move to a new `gossamer.codec`, which takes distillate — and with
it digression, inimitable and wisteria — out of the closure of every
module that merely handles text. The sources keep `package gossamer`,
so no call site is repackaged, and the four `enumIdentification` givens
already required an explicit import, so only `textDecodable`'s two
importers (both enigmatic tests) see any change.

Eight components were reaching distillate through gossamer and now
declare it: ambience, cosmopolite, geodesy, nomenclature, urticose,
legerdemain.query and probably.coverage, which also reached digression
the same way.

gossamer.core's closure falls from 26 modules to 23 and the critical
path from 25 to 24.
