+++ build.mill
@@ -726,7 +726,7 @@ object soundness extends Toolchain:
   object base extends Bundle(ambience.core, aperture.core, aviation.core, contextual.core,
     contingency.core, denominative.core, digression.core, diuretic.core, fulminate.core,
     hieroglyph.core, adversaria.core, turbulence.core, turbulence.stdio, eucalyptus.core,
-    gossamer.core, gossamer.lexicon,
+    gossamer.core, gossamer.codec, gossamer.lexicon,
     fulminate.print,
     inimitable.core, iridescence.core, kaleidoscope.core, mercator.core, nomenclature.core,
     nomenclature.lexicon,
@@ -1059,7 +1059,8 @@ object adversaria extends Library:
       settings.sep(super.scalacOptions())
 
 object ambience extends Library:
-  object core extends Component(anticipation.path, anticipation.print, gossamer.core):
+  object core extends Component(anticipation.path, anticipation.print, gossamer.core,
+      distillate.core):
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
   // WASI-backed environment: an `Environment` given whose variables come from a Wasm Component
@@ -1505,7 +1506,7 @@ object corpuscular extends Library:
       settings.sep(super.scalacOptions())
 
 object cosmopolite extends Library:
-  object core extends Component(gossamer.core):
+  object core extends Component(distillate.core, gossamer.core):
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
   object test extends Tests(core):
@@ -1963,7 +1964,7 @@ object geodesy extends Library:
   object angle extends Component(hypotenuse.core):
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
-  object core extends Component(angle, gossamer.core):
+  object core extends Component(distillate.core, angle, gossamer.core):
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
   object test extends Tests(core, angle):
@@ -1999,7 +2000,14 @@ object gnossienne extends Library:
       settings.sep(super.scalacOptions())
 
 object gossamer extends Library:
-  object core extends Component(hieroglyph.core, kaleidoscope.core, spectacular.core, hypotenuse.core, distillate.core):
+  object core extends Component(hieroglyph.core, kaleidoscope.core, spectacular.core,
+      hypotenuse.core):
+    override def scalacOptions = Task:
+      settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
+  // Decoding `Text` from bytes and naming enum cases: gossamer's only uses of distillate,
+  // and so the component that keeps distillate (with digression, inimitable and wisteria
+  // behind it) out of the closure of everything that merely handles text.
+  object codec extends Component(core, distillate.core):
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
   // Word-lookup structures — an Aho-Corasick `Dictionary` and a fuzzy-search `Lexicon` — which
@@ -2266,7 +2274,7 @@ object legerdemain extends Library:
   // The `Query` type and its `Parametric` typeclass: the form-encoded half of the library, with
   // no HTML in it. `telekinesis` needs only this, so keeping it out of `core` takes the whole
   // HTML/XML stack off the HTTP-client path.
-  object query extends Component(gossamer.core):
+  object query extends Component(gossamer.core, distillate.core):
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
   object core extends Component(honeycomb.core, anamnesis.core, query):
@@ -2354,7 +2362,7 @@ object mosquito extends Library:
       settings.sep(super.scalacOptions())
 
 object urticose extends Library:
-  object core extends Component(gossamer.core):
+  object core extends Component(distillate.core, gossamer.core):
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
   object url extends Component(anticipation.url, serpentine.core, contextual.core, core):
@@ -2369,7 +2377,7 @@ object urticose extends Library:
       settings.sep(super.scalacOptions())
 
 object nomenclature extends Library:
-  object core extends Component(gossamer.core):
+  object core extends Component(gossamer.core, distillate.core):
     def compileMvnDeps = Seq(mvn"org.scala-lang:scala3-compiler_3:${scalaVersion()}")
     override def scalacOptions = Task:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
@@ -2567,7 +2575,7 @@ object probably extends Library:
       settings.sep(super.scalacOptions() ++ Seq("-Yno-predef", "-Yimports:java.lang,proscenium"))
   // The scoverage `Invoker`'s initializer needs `SecureRandom` (absent on Scala Native), so
   // the measurement-file lookup is split by platform (`coverage-jvm`/`coverage-native`).
-  object coverage extends Component(dendrology.tree):
+  object coverage extends Component(distillate.core, digression.core, dendrology.tree):
     override def sources = Task.Sources(moduleDir, moduleDir / os.up / "coverage-jvm")
     override def nativeSources = Task.Sources(moduleDir, moduleDir / os.up / "coverage-native")
     override def scalacOptions = Task:
